+++ lnd-grpc-client/lnd-grpc-client-autoconfigure/src/main/java/org/tbk/lightning/lnd/grpc/actuator/health/LndHealthIndicator.java
@@ -8,8 +8,8 @@ import org.lightningj.lnd.wrapper.SynchronousLndAPI;
 import org.lightningj.lnd.wrapper.ValidationException;
 import org.lightningj.lnd.wrapper.message.Chain;
 import org.lightningj.lnd.wrapper.message.GetInfoResponse;
-import org.springframework.boot.actuate.health.AbstractHealthIndicator;
-import org.springframework.boot.actuate.health.Health;
+import org.springframework.boot.health.contributor.AbstractHealthIndicator;
+import org.springframework.boot.health.contributor.Health;
 
 import java.util.Collection;
 import java.util.Collections;