+++ changelog.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Breaking
 - electrum: moved electrum RPC related code to distinct package
 - electrum: compatibility wit electrum v4.6.0
-- electrum: upgrade electrum testcontainer from v3.3.8 to v4.7.2
+- electrum: upgrade electrum testcontainer from v3.3.8 to v4.8.1
+- upgrade: update spring-boot from v3.4.4 to v4.1.1
 
 ### Changed
 - upgrade: update tor testcontainer from v0.4.7.8 to v0.4.8.10
@@ -16,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - upgrade: update docker getumbrel/btc-rpc-explorer from v3.3.0 to v3.5.0
 - upgrade: update testcontainers from v1.20.4 to v2.0.2
 - upgrade: update xchange from v5.1.0 to v5.2.2
-- upgrade: update spring-boot from v3.4.4 to v3.5.8
+- upgrade: update spring-boot from v3.4.4 to v3.5.9
 - upgrade: update lombok from v1.18.30 to v1.18.42
 - upgrade: update sqlite from v3.48.0.0 to v3.51.1.0
 - upgrade: update zxing v3.5.3 to v3.5.4